requirements.txt
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.5.32
botocore==1.8.46
bz2file==0.98
certifi==2018.1.18
chardet==3.0.4
docutils==0.14
gensim==3.3.0
idna==2.6
jmespath==0.9.3
mecab-python3==0.7
numpy==1.14.0
python-dateutil==2.6.1
requests==2.18.4
s3transfer==0.1.13
scipy==1.0.0
six==1.11.0
smart-open==1.5.6
urllib3==1.22
